src/processing/ai_processor.py
import os
import logging
import whisper

# Set up logging
logging.basicConfig(level=logging.DEBUG, filename='appcutshort.log', filemode='a',
                    format='%(asctime)s - %(levelname)s - %(message)s')

def generate_subtitles(video_path, language, progress_callback=None):
    """
    Generate subtitles for a video using Whisper.

    Args:
        video_path (str): Path to the video file or stream URL.
        language (str): Language for transcription (e.g., 'English').
        progress_callback (callable, optional): Callback to update progress.

    Returns:
        list: List of subtitle dictionaries with 'start', 'end', and 'text' keys.
    """
    try:
        # Load Whisper model
        logging.info("Loading Whisper model...")
        model = whisper.load_model("base")
        if progress_callback:
            progress_callback(10)

        # Transcribe video
        logging.info(f"Transcribing video: {video_path}")
        result = model.transcribe(video_path, language=language.lower())
        if progress_callback:
            progress_callback(50)

        # Convert transcription segments to subtitle format
        subtitles = []
        for segment in result['segments']:
            start = segment['start']
            end = segment['end']
            text = segment['text'].strip()
            if text:  # Only add non-empty subtitles
                # Normalize special characters (e.g., curly quotes to straight quotes)
                text = text.replace("’", "'").replace("‘", "'").replace("“", '"').replace("”", '"')
                subtitles.append({
                    'start': start,
                    'end': end,
                    'text': text
                })

        if progress_callback:
            progress_callback(100)

        logging.info(f"Generated {len(subtitles)} subtitle entries")
        return subtitles

    except Exception as e:
        logging.error(f"Error generating subtitles: {str(e)}")
        raise

def translate_subtitles(subtitles, target_language):
    """
    Translate subtitles to the target language (placeholder).

    Args:
        subtitles (list): List of subtitle dictionaries.
        target_language (str): Target language for translation.

    Returns:
        list: Translated subtitles.
    """
    # Placeholder for translation logic
    logging.info(f"Translating subtitles to {target_language} (placeholder)")
    return subtitles