requirements.txt
pyinstaller
pytube
PyQt6
yt-dlp
transformers
torch
openai-whisper
torchvision
torchaudio
pysrt
Pillow
requests